pyproject.toml
[project]
name = "mychat"
version = "0.1.0"
description = "A flexible chat tool for Terminal and Web"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.0.0",
    "gradio>=4.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "typer>=0.9.0",
]

[project.scripts]
mychat = "mychat.main:app"

[tool.pytest.ini_options]
pythonpath = "src"
testpaths = ["tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
